journal-ai/pyproject.toml
[project]
name = "journal-ai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
